// src/interfaces/Res.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export default interface Res {
    data:             Data;
    partial_complete: boolean;
    is_success:       boolean;
    status_code:      number;
    meta:             Meta;
}

export interface Data {
    skus:        Skus[];
    schema:      Schema;
    header_info: HeaderInfo;
    meta_data:   MetaData;
}

export interface HeaderInfo {
    header: string;
}

export interface MetaData {
    title:       string;
    description: string;
}

export interface Schema {
    "@context":      string;
    "@type":         string;
    url:             string;
    name:            string;
    numberOfItems:   number;
    itemListElement: ItemListElement[];
}

export interface ItemListElement {
    "@type":  Type;
    position: number;
    url:      string;
    name:     string;
    image:    string;
}

export enum Type {
    ListItem = "ListItem",
}

export interface Skus {
    is_discontinued:   boolean;
    manufacturer_name: string;
    type:              TypeEnum;
    price:             number;
    name:              string;
    id:                number;
    sku_id:            number;
    available:         boolean;
    pack_size_label:   string;
    rx_required:       RxRequired | null;
    slug:              string;
    short_composition: string;
    image_url:         string;
    in_stock:          InStock | null;
    quantity:          number;
}

export interface InStock {
    add:          Add;
    cart_actions: CartActions;
}

export interface Add {
    action_label: ActionLabel;
    final_label:  FinalLabel;
    label:        Label;
}

export enum ActionLabel {
    Adding = "Adding..",
}

export enum FinalLabel {
    Added = "ADDED",
}

export enum Label {
    Add = "ADD",
}

export interface CartActions {
    dweb: Web[];
    mweb: Web[];
}

export enum Web {
    Add = "add",
}

export interface RxRequired {
    header:   Header;
    icon_url: string;
}

export enum Header {
    PrescriptionRequired = "Prescription Required",
}

export enum TypeEnum {
    Allopathy = "allopathy",
}

export interface Meta {
    count:       number;
    total_count: number;
}